Update IntelliJ plugin installation instructions
The documentation for installing the IntelliJ plugin has been expanded for clarity. It now includes instructions on how to correctly open a Genesis project to view the tool window bars and buttons, as well as additional information on the folder structure the plugin expects to find. This should aid users in successful project setup.